more code refactoring
