+++ webviz_subsurface/plugins/_rft_plotter/_business_logic.py
@@ -1,8 +1,13 @@
-from typing import Any, Dict, List
+from pathlib import Path
+from typing import Any, Callable, Dict, List, Optional, Tuple
 
 import numpy as np
 import pandas as pd
+from webviz_config import WebvizSettings
+from webviz_config.common_cache import CACHE
+from webviz_config.webviz_store import webvizstore
 
+from webviz_subsurface._datainput.fmu_input import load_csv
 from webviz_subsurface._utils.unique_theming import unique_colors
 
 from ._processing import filter_frame
@@ -11,19 +16,58 @@ from ._processing import filter_frame
 class RftPlotterDataModel:
     def __init__(
         self,
-        formations,
-        ertdatadf,
-        simdf,
-        formationdf,
-        faultlinesdf,
-        obsdatadf,
+        webviz_settings: WebvizSettings,
+        ensembles: Optional[List[str]],
+        formations: Path = None,
+        faultlines: Path = None,
+        obsdata: Path = None,
+        csvfile_rft: Path = None,
+        csvfile_rft_ert: Path = None,
     ):
         self.formations = formations
-        self.ertdatadf = ertdatadf
-        self.simdf = simdf
-        self.formationdf = formationdf
-        self.faultlinesdf = faultlinesdf
-        self.obsdatadf = obsdatadf
+        self.faultlines = faultlines
+        self.obsdata = obsdata
+        self.csvfile_rft = csvfile_rft
+        self.csvfile_rft_ert = csvfile_rft_ert
+
+        self.simdf = read_csv(self.csvfile_rft) if csvfile_rft is not None else None
+        self.formationdf = read_csv(self.formations) if self.formations else None
+        self.faultlinesdf = read_csv(self.faultlines) if self.faultlines else None
+        self.obsdatadf = read_csv(self.obsdata) if self.obsdata else None
+        self.ertdatadf = pd.DataFrame()
+
+        if csvfile_rft_ert and ensembles:
+            raise ValueError(
+                'Incorrent arguments. Either provide a "csvfile_rft_ert" or "ensembles"'
+            )
+
+        if csvfile_rft_ert is not None:
+            self.ertdatadf = read_csv(self.csvfile_rft_ert)
+
+        if ensembles:
+            self.ens_paths = (
+                {
+                    ens: webviz_settings.shared_settings["scratch_ensembles"][ens]
+                    for ens in ensembles
+                }
+                if ensembles is not None
+                else None
+            )
+
+            try:
+                self.simdf = load_csv(self.ens_paths, "share/results/tables/rft.csv")
+            except (KeyError, OSError):
+                self.simdf = None
+
+            try:
+                self.ertdatadf = load_csv(
+                    self.ens_paths, "share/results/tables/rft_ert.csv"
+                )
+            except KeyError as exc:
+                raise KeyError(
+                    "CSV file for ERT RFT observations/simulations "
+                    "(share/results/tables/rft_ert.csv) not found!"
+                ) from exc
 
         self.ertdatadf = self.ertdatadf.rename(
             columns={
@@ -109,3 +153,57 @@ class RftPlotterDataModel:
                 },
             }
         return marks
+
+    @property
+    def webviz_store(self) -> List[Tuple[Callable, List[Dict[str, Any]]]]:
+        functions: List[Tuple[Callable, List[Dict[str, Any]]]] = [
+            (
+                read_csv,
+                [
+                    {"csv_file": path}
+                    for path in [
+                        self.faultlines,
+                        self.formations,
+                        self.obsdata,
+                        self.csvfile_rft,
+                        self.csvfile_rft_ert,
+                    ]
+                    if path is not None
+                ],
+            )
+        ]
+        if self.csvfile_rft_ert is None:
+            functions.append(
+                (
+                    load_csv,
+                    [
+                        {
+                            "ensemble_paths": self.ens_paths,
+                            "csv_file": "share/results/tables/rft_ert.csv",
+                        },
+                    ],
+                )
+            )
+            try:
+                load_csv(self.ens_paths, "share/results/tables/rft.csv")
+                functions.append(
+                    (
+                        load_csv,
+                        [
+                            {
+                                "ensemble_paths": self.ens_paths,
+                                "csv_file": "share/results/tables/rft.csv",
+                            },
+                        ],
+                    )
+                )
+            except KeyError:
+                pass
+
+        return functions
+
+
+@CACHE.memoize(timeout=CACHE.TIMEOUT)
+@webvizstore
+def read_csv(csv_file: str) -> pd.DataFrame:
+    return pd.read_csv(csv_file)
